% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMcpp}
\alias{EMcpp}
\title{Based on EM algorithm and boot algorithm, the survival analysis of censored data was carried out}
\usage{
EMcpp(x, s, y, u, v, lambda0 = 0.1, max_it = 10000L, tol = 1e-06)
}
\arguments{
\item{x}{the results of the first type of experiment}

\item{s}{cut-off values in the first type of experiment}

\item{y}{the results of the second type of experiment}

\item{u}{the results of the third type of experiment}

\item{v}{the results of the third type of experiment}

\item{lambda0}{the initial value of the iteration}

\item{max.it}{the maximum number of iterations}

\item{lambda}{parameters of the survival model}

\item{B}{number of bootstrap samples}

\item{alpha}{the level of significance of the confidence interval}
}
\value{
a list with maximum likelihood estimates for parameters, full values for censored data, and estimated expectations, variances, and confidence intervals
}
\description{
Based on EM algorithm and boot algorithm, the survival analysis of censored data was carried out
}
