\name{SA23204158-package}
\alias{SA23204158-package}
\docType{package}
\title{This software package is used to deal with data censoring in survival analysis, using the E-M algorithm to obtain the maximum likelihood estimate of the parameters, using Bayesian posterior expectations to supplement the censored values， and using the Bootstrap algorithm to compute the expected variance and confidence interval of parameter estimation.}
\description{R and Rcpp versions of E-M algorithm and Bootstrap algorithm}
\author{
Zekang Feng
Maintainer: Zekang Feng <feng_zekang@mail.ustc.edu.cn>
}
\keyword{ package }